pyproject.toml
[tool.poetry]
name = "classroom"
version = "0.1.0"
description = ""
authors = ["Neeraj Kumar <sainineeraj1234@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
djangorestframework = "^3.15.2"
django-storages = "^1.14.4"
boto3 = "^1.34.144"
django-constance = "^3.1.0"
django-extensions = "^3.2.3"
requests = "^2.32.3"
pyotp = "^2.9.0"
django-phonenumber-field = "^8.0.0"
djangorestframework-simplejwt = "^5.3.1"
phonenumbers = "^8.13.40"
pillow = "^10.4.0"
drf-yasg = "^1.21.7"
setuptools = "^72.0.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "django"

[tool.flake8]
max-line-length = 120
